make configuration a sep. module, including arguments
