+++ freqtrade/configuration/__init__.py
@@ -1,0 +1,8 @@
+from freqtrade.configuration.arguments import Arguments, TimeRange  # noqa: F401
+
+from freqtrade.configuration.arguments import (  # noqa: F401
+        ARGS_DOWNLOADER,
+        ARGS_PLOT_DATAFRAME,
+        ARGS_PLOT_PROFIT)
+
+from freqtrade.configuration.configuration import Configuration  # noqa: F401
